refactor(permissions): Rename PermissionRoles to permissions
