+++ src/auth/authoptions.ts
@@ -4,7 +4,7 @@ import { decryptAndComparePassword } from './password'
 import FeideProvider from '@/lib/feide/FeideProvider'
 import { updateUserStudyProgrammes } from '@/lib/feide/userRoutines'
 import prisma from '@/prisma'
-import { readPermissionsOfUser } from '@/services/permissionRoles/read'
+import { readPermissionsOfUser } from '@/services/permissions/read'
 import { readMembershipsOfUser } from '@/services/groups/memberships/read'
 import { updateEmailForFeideAccount } from '@/services/auth/feideAccounts/update'
 import { UserMethods } from '@/services/users/methods'
