+++ src/services/shop/purchase/methods.ts
@@ -2,7 +2,7 @@ import '@pn-server-only'
 import { PurchaseAuthers } from './authers'
 import { PurchaseSchemas } from './schemas'
 import { ServerError } from '@/services/error'
-import { readPermissionsOfUser } from '@/services/permissionRoles/read'
+import { readPermissionsOfUser } from '@/services/permissions/read'
 import { ServiceMethod } from '@/services/ServiceMethod'
 import { UserMethods } from '@/services/users/methods'
 import { UserConfig } from '@/services/users/config'
