+++ src/services/users/methods.ts
@@ -2,7 +2,7 @@ import '@pn-server-only'
 import { UserAuthers } from './authers'
 import { UserConfig } from './config'
 import { readMembershipsOfUser } from '@/services/groups/memberships/read'
-import { readPermissionsOfUser } from '@/services/permissionRoles/read'
+import { readPermissionsOfUser } from '@/services/permissions/read'
 import { NTNUEmailDomain } from '@/services/mail/mailAddressExternal/ConfigVars'
 import { sendVerifyEmail } from '@/services/notifications/email/systemMail/verifyEmail'
 import { createDefaultSubscriptions } from '@/services/notifications/subscription/create'
